Move static page views navigable from home page to own module
