+++ bagitobjecttransfer/recordtransfer/views.py
@@ -103,12 +103,6 @@ from recordtransfer.utils import (
 LOGGER = logging.getLogger(__name__)
 
 
-class Index(TemplateView):
-    """The homepage."""
-
-    template_name = "recordtransfer/home.html"
-
-
 def media_request(request: HttpRequest, path: str) -> HttpResponse:
     """Respond to whether a media request is allowed or not."""
     if not request.user.is_authenticated:
@@ -138,12 +132,6 @@ class TransferSent(TemplateView):
     """The page a user sees when they finish a transfer."""
 
     template_name = "recordtransfer/transfersent.html"
-
-
-class SystemErrorPage(TemplateView):
-    """The page a user sees when there is some system error."""
-
-    template_name = "recordtransfer/systemerror.html"
 
 
 class UserProfile(UpdateView):
@@ -244,20 +232,6 @@ class UserProfile(UpdateView):
         profile_form = cast(UserProfileForm, form)
         profile_form.reset_form()
         return super().form_invalid(profile_form)
-
-
-class About(TemplateView):
-    """About the application."""
-
-    template_name = "recordtransfer/about.html"
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        context["accepted_files"] = settings.ACCEPTED_FILE_FORMATS
-        context["max_total_upload_size"] = settings.MAX_TOTAL_UPLOAD_SIZE
-        context["max_single_upload_size"] = settings.MAX_SINGLE_UPLOAD_SIZE
-        context["max_total_upload_count"] = settings.MAX_TOTAL_UPLOAD_COUNT
-        return context
 
 
 class ActivationSent(TemplateView):
